operator: Exploit _specialize_clusters
