+++ devito/core/gpu.py
@@ -2,9 +2,13 @@ import cgen as c
 
 from devito.core.operator import OperatorCore
 from devito.data import FULL
+from devito.dse import rewrite
+from devito.ir.clusters import Toposort, analyze
 from devito.ir.support import COLLAPSED
+from devito.passes.clusters import Lift, fuse, scalarize, eliminate_arrays
 from devito.passes import (DataManager, Ompizer, ParallelTree, optimize_halospots,
                            mpiize, hoist_prodders)
+from devito.tools import generator
 
 __all__ = ['DeviceOffloadingOperator']
 
@@ -132,6 +136,39 @@ class OffloadingDataManager(DataManager):
 
 
 class DeviceOffloadingOperator(OperatorCore):
+
+    @classmethod
+    def _specialize_clusters(cls, clusters, **kwargs):
+        # TODO: this is currently identical to CPU64NoopOperator._specialize_clusters,
+        # but it will have to change
+
+        mode = kwargs['dse']
+
+        # To create temporaries
+        counter = generator()
+        template = lambda: "r%d" % counter()
+
+        # Toposort+Fusion (the former to expose more fusion opportunities)
+        clusters = Toposort().process(clusters)
+        clusters = fuse(clusters)
+
+        # Flop reduction via the DSE
+        clusters = rewrite(clusters, template, mode=mode)
+
+        # Lifting
+        clusters = Lift().process(clusters)
+
+        # Lifting may create fusion opportunities, which in turn may enable
+        # further optimizations
+        clusters = fuse(clusters)
+        clusters = eliminate_arrays(clusters, template)
+        clusters = scalarize(clusters, template)
+
+        # Determine computational properties (e.g., parallelism) that will be
+        # used by the later passes
+        clusters = analyze(clusters)
+
+        return clusters
 
     @classmethod
     def _specialize_iet(cls, graph, **kwargs):
