+++ devito/ir/clusters/__init__.py
@@ -1,2 +1,4 @@
+from .queue import *  # noqa
+from .analysis import *  # noqa
 from .cluster import *  # noqa
 from .algorithms import *  # noqa
