+++ devito/ops/operator.py
@@ -1,4 +1,6 @@
 from devito import Eq
+from devito.dse import rewrite
+from devito.ir.clusters import Toposort, analyze
 from devito.ir.equations import ClusterizedEq
 from devito.ir.iet import Call, Expression, find_affine_trees
 from devito.ir.iet.visitors import FindSymbols, Transformer
@@ -8,9 +10,10 @@ from devito.ops import ops_configuration
 from devito.ops.types import OpsBlock
 from devito.ops.transformer import create_ops_dat, create_ops_fetch, opsit
 from devito.ops.utils import namespace
-from devito.passes import DataManager, iet_pass
+from devito.passes.clusters import Lift, fuse, scalarize, eliminate_arrays
+from devito.passes.iet import DataManager, iet_pass
 from devito.symbolics import Literal
-from devito.tools import filter_sorted, flatten
+from devito.tools import filter_sorted, flatten, generator
 
 __all__ = ['OPSOperator']
 
@@ -22,6 +25,39 @@ class OPSOperator(Operator):
     """
 
     _default_headers = Operator._default_headers + ['#define restrict __restrict']
+
+    @classmethod
+    def _specialize_clusters(cls, clusters, **kwargs):
+        # TODO: this is currently identical to CPU64NoopOperator._specialize_clusters,
+        # but it will have to change
+
+        mode = kwargs['dse']
+
+        # To create temporaries
+        counter = generator()
+        template = lambda: "r%d" % counter()
+
+        # Toposort+Fusion (the former to expose more fusion opportunities)
+        clusters = Toposort().process(clusters)
+        clusters = fuse(clusters)
+
+        # Flop reduction via the DSE
+        clusters = rewrite(clusters, template, mode=mode)
+
+        # Lifting
+        clusters = Lift().process(clusters)
+
+        # Lifting may create fusion opportunities, which in turn may enable
+        # further optimizations
+        clusters = fuse(clusters)
+        clusters = eliminate_arrays(clusters, template)
+        clusters = scalarize(clusters, template)
+
+        # Determine computational properties (e.g., parallelism) that will be
+        # used by the later passes
+        clusters = analyze(clusters)
+
+        return clusters
 
     @classmethod
     def _specialize_iet(cls, graph, **kwargs):
