+++ devito/passes/__init__.py
@@ -1,1 +1,2 @@
+from .clusters import *  # noqa
 from .iet import *  # noqa
