+++ devito/passes/clusters/__init__.py
@@ -1,0 +1,1 @@
+from .misc import *  # noqa
